:art: Set grub options via SDK
